Add a server script to launch the app
I've been playing with various technologies recently and remembering how to start up the various projects has become a challenge so decided to include the server script that is part of the standard github scripts-to-rule-them-all.
In this case it just runs update to ensure everything is up to date and highlights where to add additional commands for a specific project.